Move prover CLI to ethrex CLI
