+++ cmd/ethrex/Cargo.toml
@@ -49,6 +49,7 @@ cfg-if = "1.0.0"
 ethrex-dev = { path = "../../crates/blockchain/dev", optional = true }
 ethrex-metrics = { path = "../../crates/blockchain/metrics" }
 url.workspace = true
+ethrex-prover = { workspace = true, features = ["l2"] }
 
 [[bin]]
 name = "ethrex"
@@ -76,6 +77,8 @@ rollup_storage_libmdbx = ["ethrex-storage-rollup/libmdbx"]
 rollup_storage_redb = ["ethrex-storage-rollup/redb"]
 rollup_storage_sql = ["ethrex-storage-rollup/sql"]
 sync-test = []
+sp1 = ["ethrex-prover/sp1"]
+
 [dev-dependencies]
 criterion = { version = "0.5.1", features = [
   "html_reports",
