+++ crates/l2/prover/Cargo.toml
@@ -47,10 +47,6 @@ ethrex-storage.workspace = true
 name = "ethrex_prover_lib"
 path = "src/lib.rs"
 
-[[bin]]
-name = "ethrex_prover"
-path = "src/main.rs"
-
 [features]
 # TODO(#2662) Prover shouldn't have l2 as a default but cargo complains about missing field in struct
 default = ["l2"]
